test: move binding entry tests to compiler cases
